Two sample assignments from 2016 Data Science course
In the 'estimation' assignment, data about the 2012 presidential
election voting share was combined with State of the Union addresses and
a Laplace-smoothed bigram model of language to produce a model capable
of both giving the log probability that a given sentence came from a
Republican or Democrat, as well as producing a random (and probably
completely unintelligible--the bigram model really isn't THAT good)
sentence representative of either party.

In the open-ended 'regression' assignment, CSV data of my choosing was
pulled from the internet before the 2016 election in an attempt to
create the most accurate prediction of the election outcome. I used a
2nd degree polynomial fit for the regression, along with data indicating
population density, percentage of population with advanced degrees,
standard income tax rate, and minority percentage of population (all per
state). all of which led--whether by deeply ingrained intuitive
foresight of the micro and macro political shifts of the entire United
States, strengthened to infallability by a whopping ten weeks of
statistical study, or by pretty much chance and a basic understanding of
correlation, you decide--to a prediction more accurate for this election
than that of famous blog FiveThirtyEight.